chore(iast): reorganize tests
