+++ tests/appsec/integrations/flask_tests/test_iast_flask_testagent.py
@@ -7,34 +7,9 @@ from ddtrace.appsec._iast.constants import VULN_CODE_INJECTION
 from ddtrace.appsec._iast.constants import VULN_HEADER_INJECTION
 from ddtrace.appsec._iast.constants import VULN_STACKTRACE_LEAK
 from tests.appsec.appsec_utils import flask_server
-from tests.appsec.integrations.flask_tests.test_flask_remoteconfig import _get_agent_client
-
-
-def start_trace(token):
-    client = _get_agent_client()
-    client.request(
-        "GET", "/test/session/start?test_session_token=%s" % (token,), headers={"X-Datadog-Test-Session-Token": token}
-    )
-    resp = client.getresponse()
-    return resp.read()
-
-
-def clear_session(token):
-    client = _get_agent_client()
-    client.request(
-        "GET", "/test/session/clear?test_session_token=%s" % (token,), headers={"X-Datadog-Test-Session-Token": token}
-    )
-    resp = client.getresponse()
-    return resp.read()
-
-
-def _get_span(token):
-    client = _get_agent_client()
-    client.request(
-        "GET", "/test/session/traces?test_session_token=%s" % (token,), headers={"X-Datadog-Test-Session-Token": token}
-    )
-    resp = client.getresponse()
-    return json.loads(resp.read())
+from tests.appsec.integrations.utils_testagent import _get_span
+from tests.appsec.integrations.utils_testagent import clear_session
+from tests.appsec.integrations.utils_testagent import start_trace
 
 
 @pytest.mark.skip(reason="Stacktrace error in debug mode doesn't work outside the request APPSEC-56862")
